pyproject.toml
[project]
name = "justhost-assignment"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.5",
    "django-environ>=0.12.0",
    "django-filter>=24.3",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "pytest-factoryboy>=2.7.0",
    "ruff>=0.9.1",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
extend-select = [
    "E",
    "F",
    "DJ",
    "I",
    "ANN",
]
ignore = ["ANN204", "ANN002", "ANN003"]
dummy-variable-rgx = "__.*"
exclude = ["manage.py", "vps/migrations/*.py"]

[tool.ruff.format]
quote-style = "double"
exclude = ["manage.py", "vps/migrations/*.py"]
